Backend controllers and services refactor

+++ src/main/java/com/podium/controller/ContactController.java
@@ -3,11 +3,9 @@ package com.podium.controller;
 import com.podium.constant.PodiumEndpoint;
 import com.podium.model.dto.request.ContactRequestDto;
 import com.podium.model.dto.response.ContactResponseDto;
-import com.podium.model.dto.response.SubjectResponseDto;
-import com.podium.model.dto.request.SubjectRequestDto;
 import com.podium.service.ContactService;
+import com.podium.service.SubjectService;
 import com.podium.validation.main.PodiumValidator;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -18,10 +16,11 @@ import org.springframework.web.server.ResponseStatusException;
 public class ContactController {
 
     private ContactService contactService;
+    private SubjectService subjectService;
 
-    @Autowired
-    public ContactController(ContactService contactService) {
+    public ContactController(ContactService contactService, SubjectService subjectService) {
         this.contactService = contactService;
+        this.subjectService = subjectService;
     }
 
     @PostMapping(PodiumEndpoint.addContact)
@@ -29,7 +28,7 @@ public class ContactController {
 
         PodiumValidator.getInstance().validateRequestBody(request);
 
-        if(!this.contactService.existSubjectByName(request.getSubject()))
+        if(!this.subjectService.existSubjectByName(request.getSubject()))
             throw new ResponseStatusException(
                     HttpStatus.CONFLICT, "Given subject does not exist");
 
@@ -42,12 +41,19 @@ public class ContactController {
     @DeleteMapping(PodiumEndpoint.deleteContact)
     public ResponseEntity deleteContact(@PathVariable int id){
 
-        this.contactService.deleteContact(id);
-        return ResponseEntity.ok().body("Contact deleted");
+        if(this.contactService.existContactById(id)){
+            this.contactService.deleteContact(id);
+            return ResponseEntity.ok().body("Contact deleted");
+        }
+
+        else throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Contact with given id does not exist");
+
     }
 
     @GetMapping(PodiumEndpoint.findAllContactByEmail)
-    public ResponseEntity findContactByEmail(@PathVariable String email){
+    public ResponseEntity<Iterable<ContactResponseDto>>
+    findAllContactByEmail(@PathVariable String email){
 
         return ResponseEntity
                 .ok()
@@ -55,63 +61,20 @@ public class ContactController {
     }
 
     @GetMapping(PodiumEndpoint.findAllContactBySubject)
-    public ResponseEntity findContactBySubject(@PathVariable String subject){
+    public ResponseEntity<Iterable<ContactResponseDto>>
+    findAllContactBySubject(@PathVariable String subject){
 
         return ResponseEntity
                 .ok()
                 .body(this.contactService.findAllBySubject(subject));
     }
 
-    @PostMapping(PodiumEndpoint.addSubject)
-    public ResponseEntity addSubject(@RequestBody SubjectRequestDto request){
-
-        PodiumValidator.getInstance().validateRequestBody(request);
-
-        if(this.contactService.existSubjectByName(request.getSubject()))
-            throw new ResponseStatusException(
-                    HttpStatus.CONFLICT, "Subject already exists");
-
-        this.contactService.addSubject(request.getSubject());
-        return ResponseEntity.ok().body("Subject successfully added");
-
-    }
-
-    @DeleteMapping(PodiumEndpoint.deleteSubject)
-    public ResponseEntity deleteSubject(@PathVariable String name){
-
-        this.contactService.deleteSubjectByName(name);
-        return ResponseEntity.ok().body("Subject deleted");
-    }
-
-    @GetMapping(PodiumEndpoint.findAllSubject)
-    public Iterable<SubjectResponseDto> findAllSubjects(){
-        return this.contactService.findAllSubjects();
-    }
-
     @GetMapping(PodiumEndpoint.findAllContact)
-    public Iterable<ContactResponseDto> findAllContact(){
-        return this.contactService.findAllContact();
-    }
-
-    @GetMapping("/subject/find/{name}")
-    public ResponseEntity findSubjectByName(@PathVariable String name){
+    public ResponseEntity<Iterable<ContactResponseDto>>
+    findAllContact(){
         return ResponseEntity
                 .ok()
-                .body(this.contactService.findSubjectByName(name));
+                .body(this.contactService.findAllContact());
     }
-
-    @GetMapping(PodiumEndpoint.existSubjectByName)
-    public ResponseEntity existSubjectByName(@PathVariable String name){
-
-        if(this.contactService.existSubjectByName(name))
-            return ResponseEntity.ok().build();
-
-        else
-            return ResponseEntity.notFound().build();
-
-    }
-
-
-
 
 }
